Moved the code for adding an attachment to a separate trait.
