+++ src/Picqer/Financials/Moneybird/Entities/ExternalSalesInvoice.php
@@ -2,6 +2,7 @@
 
 namespace Picqer\Financials\Moneybird\Entities;
 
+use Picqer\Financials\Moneybird\Actions\Attachment;
 use Picqer\Financials\Moneybird\Actions\Downloadable;
 use Picqer\Financials\Moneybird\Actions\Filterable;
 use Picqer\Financials\Moneybird\Actions\FindAll;
@@ -9,6 +10,7 @@ use Picqer\Financials\Moneybird\Actions\FindOne;
 use Picqer\Financials\Moneybird\Actions\Removable;
 use Picqer\Financials\Moneybird\Actions\Storable;
 use Picqer\Financials\Moneybird\Actions\Synchronizable;
+use Picqer\Financials\Moneybird\Connection;
 use Picqer\Financials\Moneybird\Model;
 
 /**
@@ -19,7 +21,7 @@ use Picqer\Financials\Moneybird\Model;
  */
 class ExternalSalesInvoice extends Model
 {
-    use FindAll, FindOne, Storable, Removable, Filterable, Downloadable, Synchronizable;
+    use FindAll, FindOne, Storable, Removable, Filterable, Downloadable, Synchronizable, Attachment;
 
     /**
      * @var array
@@ -83,4 +85,11 @@ class ExternalSalesInvoice extends Model
             'type' => self::NESTING_TYPE_ARRAY_OF_OBJECTS,
         ],
     ];
+
+    public function __construct(Connection $connection, array $attributes = [])
+    {
+        parent::__construct($connection, $attributes);
+
+        $this->attachmentPath = 'attachment';
+    }
 }
