+++ src/Picqer/Financials/Moneybird/Entities/PurchaseInvoice.php
@@ -2,7 +2,7 @@
 
 namespace Picqer\Financials\Moneybird\Entities;
 
-use Picqer\Financials\Moneybird\Model;
+use Picqer\Financials\Moneybird\Actions\Attachment;
 use Picqer\Financials\Moneybird\Actions\FindAll;
 use Picqer\Financials\Moneybird\Actions\FindOne;
 use Picqer\Financials\Moneybird\Actions\Storable;
@@ -10,13 +10,14 @@ use Picqer\Financials\Moneybird\Actions\Removable;
 use Picqer\Financials\Moneybird\Actions\Filterable;
 use Picqer\Financials\Moneybird\Actions\Synchronizable;
 use Picqer\Financials\Moneybird\Exceptions\ApiException;
+use Picqer\Financials\Moneybird\Model;
 
 /**
  * Class PurchaseInvoice.
  */
 class PurchaseInvoice extends Model
 {
-    use FindAll, FindOne, Storable, Removable, Filterable, Synchronizable;
+    use FindAll, FindOne, Storable, Removable, Filterable, Synchronizable, Attachment;
 
     /**
      * @var array
