+++ src/Picqer/Financials/Moneybird/Entities/SalesInvoice.php
@@ -3,6 +3,7 @@
 namespace Picqer\Financials\Moneybird\Entities;
 
 use InvalidArgumentException;
+use Picqer\Financials\Moneybird\Actions\Attachment;
 use Picqer\Financials\Moneybird\Actions\Downloadable;
 use Picqer\Financials\Moneybird\Actions\Filterable;
 use Picqer\Financials\Moneybird\Actions\FindAll;
@@ -22,7 +23,7 @@ use Picqer\Financials\Moneybird\Model;
  */
 class SalesInvoice extends Model
 {
-    use FindAll, FindOne, Storable, Removable, Filterable, Downloadable, Synchronizable;
+    use FindAll, FindOne, Storable, Removable, Filterable, Downloadable, Synchronizable, Attachment;
 
     /**
      * @var array
@@ -260,33 +261,6 @@ class SalesInvoice extends Model
         );
 
         return $this->makeFromResponse($response);
-    }
-
-    /**
-     * Add Attachment to this invoice.
-     *
-     * You can use fopen('/path/to/file', 'r') in $resource.
-     *
-     * @param string $filename The filename of the attachment
-     * @param resource $contents A StreamInterface/resource/string, @see http://docs.guzzlephp.org/en/stable/request-options.html?highlight=multipart#multipart
-     *
-     * @return \Picqer\Financials\Moneybird\Entities\SalesInvoice
-     *
-     * @throws \Picqer\Financials\Moneybird\Exceptions\ApiException
-     */
-    public function addAttachment($filename, $contents)
-    {
-        $this->connection()->upload($this->endpoint . '/' . $this->id . '/attachments', [
-            'multipart' => [
-                [
-                    'name' => 'file',
-                    'contents' => $contents,
-                    'filename' => $filename,
-                ],
-            ],
-        ]);
-
-        return $this;
     }
 
     /**
